Add section two, styling and positionning
Section Two wasn't hard to build, using the card method. I thought it
would be much harder because of the images inside the cards but because
I've defined the width of the cards, the image inside its holder take
the full width (I tought it would overwhelm the holder but apparently
not). Only thing is if you want a portrait img it will change the
overall alignment of the cards so I simply took landscape ones.