Add full section three
It really isn't hard, only to divs with some styling and positionning.
I keep forgeting that margin only applies to the same level of the
margin's element so it doesn't push away a bigger element. So I have to
use padding in the bigger element (here section-three).
I also used text-weight: lighter wich did the trick for the quote.